Cart getter refactored
