+++ saleor/registration/views.py
@@ -19,7 +19,7 @@ from django.utils.translation import ugettext_lazy as _
 from . import forms
 from .models import EmailConfirmationRequest, EmailChangeRequest
 from . import utils
-from ..cart.decorators import assign_anonymous_cart
+from ..cart.views import assign_anonymous_cart
 
 now = timezone.now
